Refactored test packages
