+++ orienteer-core/src/test/resources/META-INF/services/com.google.inject.Module
@@ -1,3 +1,3 @@
 org.orienteer.junit.OrienteerTestModule
 org.orienteer.core.service.CustomTestModule
-org.orienteer.core.distributed.boot.loader.service.DistributedTestInitModule
+org.orienteer.core.boot.loader.distributed.service.DistributedTestInitModule
