+++ orienteer-core/src/test/resources/org/orienteer/core/boot/loader/distributed/modules.xml
