+++ orienteer-core/src/test/resources/org/orienteer/core/boot/loader/distributed/orienteer-pages.jar
